Fix margin parsing.
PWG JPS3 [1] says margins are given in 1/100 of mm instead of mm as
expected by the previous code. This mismatch can cause the margins to be
larger than the page: e.g. for my HP LaserJet 1018, the margins are 4 mm,
thus the previous code calculated margins of 40 cm on all sides, which
is larger than the 210x297 mm A4 page.

Should fix #7.

[1] https://ftp.pwg.org/pub/pwg/candidates/cs-ippjobprinterext3v10-20120727-5100.13.pdf