better detection of release track
When publishing to the play store, the environment variable PLAY_STORE_RELEASE_CHANNEL is used to determine to which track the apk should be pushed.
When debug building, this has no meaning and the build should pass even if this environment variable is not defined.
So the new behavior is to set the track to what's defined in the environment variable only if it is defined.