Filter for CUPS-Get-Printers Request
If we want to get only one printer, we still request the complete list of printers from the cups server. But in the "CUPS-Get-Printers Request" we could define a "first-printer-name" and a "limit". With these attributes it is possible to get only the wanted printer and not the whole list of printers. If the list of printers is very long, this request should be much faster.