NullPointerException in updateJobStatus
However my CupsServer don't switch the status into completed. After some time it is gone and updateJobStatus throws a NullPointerException, because it don't find the job. So then we should also complete the printJob. And of course at all Exceptions we should stop the polling. Otherwise we run again and again into the Exception. At the other Exceptions, i insert the error message of the Android PrintingService.