Stop the PrinterDiscovery and PrinterStateTracking
Make the IppOperation abortable, like the original cups4j. With this it is possible to stop the PrinterStateTracking if you change the printer or cancel the printing Dialog. With this the network performance should be better, because there isn't unnecessary traffic. Also stop the PrinterDiscovery now, if you cancel the printing Dialog. We shouldn't search for printers if we don't need it any more.